[S05P13C104-35]/FE_RoomUI : Add back + front

+++ backend/src/main/java/com/ssafy/api/controller/ConferenceController.java
@@ -168,7 +168,7 @@ public class ConferenceController {
         String userId = userDetails.getUsername();
 
         registerPostReq.setUserid(userId);
-        Long result = conferenceService.register(registerPostReq);
+        int result = conferenceService.register(registerPostReq);
 
         return ResponseEntity.ok(ConferenceRes.of(201,"Success",result));
